+++ CHANGELOG.md
@@ -107,3 +107,5 @@
 * Define `<.:>`, `>.:>`, `<.:<`, `>.:<` type aliases
 * Change `Casting` typeclass definition
 * Define `extract` method
+* Define `absurd` method
+* Define `empty` method